aider: Changed grid layout to a single column for viewports with a width of less than 800 pixels.
# Aider chat conversation:

ASSISTANT: To achieve this, we need to modify the media query in the `pages/index.js` file to change the grid layout for viewports with a width of less than 800 pixels. We will update the CSS for the `.grid` class within the media query to use a single column layout.

Here is the *SEARCH/REPLACE* block for the change:

pages/index.js
```jsx
<<<<<<< SEARCH
        @media (max-width: 600px) {
          .grid {
            width: 100%;
            flex-direction: column;
          }
        }
=======
        @media (max-width: 800px) {
          .grid {
            grid-template-columns: 1fr; /* Change to single column layout */
          }
        }
>>>>>>> REPLACE
```