aider: Adjusted container and main styles to pin the footer to the bottom of the page.
# Aider chat conversation:

USER: Very good! Okay some additional styling we need is the footer is no longer pinned to the bottom of the page

ASSISTANT: To pin the footer to the bottom of the page, we can use a common CSS technique that involves setting the `min-height` of the main content area to `100vh` (viewport height) and using flexbox to push the footer to the bottom. We'll adjust the `.container` class to use flexbox and set the `flex-direction` to `column`, then make sure the `main` content grows to take available space, pushing the `footer` to the bottom.

Here are the changes needed in the `styles/app.sass` file:

styles/app.sass
```sass
<<<<<<< SEARCH
.container
  min-height: 100vh
  padding: 0 0.5rem
  display: flex
  flex-direction: column
  justify-content: center
  align-items: center
=======
.container
  min-height: 100vh
  padding: 0 0.5rem
  display: flex
  flex-direction: column
  justify-content: space-between
  align-items: center
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
// ... (continue adding any other styles that were in index.js)
=======
main
  flex: 1
>>>>>>> REPLACE
```

These changes will ensure that the `main` element takes up all available space, which will push the `footer` to the bottom of the viewport. If there's more content than fits on the screen, the `footer` will be pushed down accordingly, appearing below the content as expected.